remove tabs and combine the configs
