--- src/care-and-treatment/tabs/allergies-config.json
@@ -1,40 +1,0 @@
-{
-  "menuId": "allergies",
-  "cardTitle": "Allergies",
-  "columns": [
-    {
-      "id": "encounterDate",
-      "isDate": true,
-      "title": "Encounter Date",
-      "encounterTypes": [
-        "allergyEncounterType"
-      ],
-      "concept": "encounterDate"
-    },
-    {
-      "id": "suspectedDrug",
-      "title": "Suspected Drug",
-      "encounterTypes": [
-        "allergyEncounterType"
-      ],
-      "concept": "suspectedDrugConcept"
-    },
-    
-    {
-      "id": "drugEffect",
-      "title": "Drug Effect",
-      "encounterTypes": [
-        "allergyEncounterType"
-      ],
-      "concept": "drugEffectConcept"
-    },
-    {
-      "id": "actionTaken",
-      "title": "Action Taken",
-      "encounterTypes": [
-        "allergyEncounterType"
-      ],
-      "concept": "actionTakenConcept"
-    }
-  ]
-}
