--- src/care-and-treatment/tabs/hospitalization-config.json
@@ -1,39 +1,0 @@
-{
-  "menuId": "hospitalization",
-  "cardTitle": "Hospitalization",
-  "columns": [
-    {
-      "id": "encounterDate",
-      "isDate": true,
-      "title": "Encounter Date",
-      "encounterTypes": [
-        "hospitalizationEncounterType"
-      ],
-      "concept": "encounterDate"
-    },
-    {
-      "id": "reasonForHospitalization",
-      "title": "Reason For Hospitalization",
-      "encounterTypes": [
-        "hospitalizationEncounterType"
-      ],
-      "concept": "reasonForHospitalizationConcept"
-    },
-    {
-      "id": "duration",
-      "title": "Duration",
-      "encounterTypes": [
-        "hospitalizationEncounterType"
-      ],
-      "concept": "durationConcept"
-    },
-    {
-      "id": "durationUnit",
-      "title": "Duration Unit",
-      "encounterTypes": [
-        "hospitalizationEncounterType"
-      ],
-      "concept": "durationUnitConcept"
-    }
-  ]
-}
