--- src/care-and-treatment/tabs/image-config.json
@@ -1,23 +1,0 @@
-{
-  "menuId": "opportunisticInfections",
-  "cardTitle": "Opportunistic Infections",
-  "columns": [
-    {
-      "id": "encounterDate",
-      "isDate": true,
-      "title": "Encounter Date",
-      "encounterTypes": [
-        "imageEncounterType"
-      ],
-      "concept": "encounterDate"
-    },
-    {
-      "id": "imageOrdered",
-      "title": "Image ordered",
-      "encounterTypes": [
-      "imageEncounterType"
-      ],
-      "concept": "imageOrderedConcept"
-    }
-  ]
-}
