--- src/care-and-treatment/tabs/oi-config.json
@@ -1,39 +1,0 @@
-{
-  "menuId": "opportunisticInfections",
-  "cardTitle": "Opportunistic Infections",
-  "columns": [
-    {
-      "id": "encounterDate",
-      "isDate": true,
-      "title": "Encounter Date",
-      "encounterTypes": [
-        "oiEncounterType"
-      ],
-      "concept": "encounterDate"
-    },
-    {
-      "id": "chronicCareDx",
-      "title": "Suspected Drug",
-      "encounterTypes": [
-        "oiEncounterType"
-      ],
-      "concept": "chronicCareDxConcept"
-    },
-    {
-      "id": "whoStage",
-      "title": "WHO Stage",
-      "encounterTypes": [
-        "oiEncounterType"
-      ],
-      "concept": "whoStageConcept"
-    },
-    {
-      "id": "criteria",
-      "title": "Criteria",
-      "encounterTypes": [
-        "oiEncounterType"
-      ],
-      "concept": "criteriaConcept"
-    }
-  ]
-}
